change length to int instead of float

+++ scripts/linux/main.py
@@ -54,7 +54,7 @@ def get_macos_song_data(server):
 	metadata_lines = metadata.split('\n')
 
 	song_data = {
-		'length': int(float(metadata_lines[0]) * 1000),
+		'length': int(int(metadata_lines[0]) * 1000),
 		'album': metadata_lines[1],
 		'artist': metadata_lines[2],
 		'title': metadata_lines[3],
